Move functions to Inventory Utilities
